pyproject.toml
[project]
name = "add-subtitles-to-videos"
version = "0.1.0"
description = "Cross-platform desktop tool for generating and burning subtitles into videos."
requires-python = ">=3.12"
dependencies = [
    "imageio-ffmpeg>=0.6.0",
    "openai-whisper>=20250625",
    "PySide6>=6.9.0",
    "torch @ https://download.pytorch.org/whl/cu124/torch-2.6.0%2Bcu124-cp312-cp312-win_amd64.whl ; platform_system == 'Windows' and platform_machine == 'AMD64'",
]

[project.scripts]
add-subtitles-to-videos = "add_subtitles_to_videos:main"

[dependency-groups]
dev = [
    "pillow>=12.2.0",
    "pyinstaller>=6.19.0",
    "pytest>=8.4.0",
]

[build-system]
requires = ["uv_build>=0.11.6,<0.12.0"]
build-backend = "uv_build"

[tool.pytest.ini_options]
testpaths = ["tests"]
